fix the revert scripts too
funny how thinking clearly is so hard when using autogenerated code.

In this case, I want to revert by flipping the primary key back to the
original version.  If the revert fails, then oh well...If there is
actually data in the table, then I really don't want to be reverting and
that is the only case I can think of in which the revert will fail.